requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.0.0
plotly>=5.0.0
requests>=2.31.0
pandas-ta>=0.3.14b
gspread>=5.12.0
google-auth>=2.23.0
google-api-python-client>=2.0.0
